Added Le and Pr; Fixing flake8
